allow definition of more than 32k internal jobs in jobExecutionQueue table (smallint to int change)

+++ source/common/stored-procedures/dbo.usp_jobQueueExecute.sql
@@ -46,8 +46,8 @@ DECLARE   @projectID					[smallint]
 		, @configMaxSQLJobsRunning		[smallint]
 		, @configMaxQueueExecutionTime	[smallint]
 		, @runningJobs					[smallint]
-		, @executedJobs					[smallint]
-		, @jobQueueCount				[smallint]
+		, @executedJobs					[int]
+		, @jobQueueCount				[int]
 			
 		, @strMessage					[varchar](8000)	
 		, @currentRunning				[int]
